connect views
